rename: 페이지 추가로 인한 컴포넌트 이름 변경
