+++ package.json
@@ -13,8 +13,10 @@
     "chromatic": "npx chromatic --project-token=chpt_176e75dc7f1a514"
   },
   "dependencies": {
+    "@types/react-router-dom": "^5.3.3",
     "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.23.0"
   },
   "devDependencies": {
     "@chromatic-com/storybook": "^1.3.3",
